requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
langsmith>=0.0.77

# Azure Services
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
openai>=1.12.0

# PDF Processing
pymupdf>=1.23.0

# Background Jobs
celery>=5.3.0
redis>=5.0.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
tiktoken>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0